import Variable as necessary when doing getitem
